meta-evb: meta-evb-arm: Support Power Control operation
Following with phosphor-state-manager design to implement
chassis/host state transition and script-based power operation.

The implementation only support simple power on and off. Graceful
shutdown not support and act as force power off.

Power operation based on the files on shared folder as below.

  * /tmp/hostsharedir/power to trigger power on/off
  * /tmp/hostsharedir/pgood to notify power good status

Tested with ipmitool
  * ipmitool -I lanplus -C 17 -N 3 -p 623 -U root -P 0penBmc -H
    192.168.122.10 power on
  * ipmitool -I lanplus -C 17 -N 3 -p 623 -U root -P 0penBmc -H
    192.168.122.10 power off
  * ipmitool -I lanplus -C 17 -N 3 -p 623 -U root -P 0penBmc -H
    192.168.122.10 power soft

Tested with obmcutil on OpenBMC shell
  * obmcutil poweron
  * obmcutil poweroff

Signed-off-by: John Chung <john.chung@arm.com>
Change-Id: I7dda438ddcb49f45dee312816807602974af77c6